+++ interface/src/lib.rs
@@ -8,28 +8,6 @@ pub mod instruction;
 #[cfg(target_arch = "wasm32")]
 mod wasm;
 
-#[cfg(feature = "bincode")]
-use solana_pubkey::Pubkey;
-
-// Inline some constants to avoid dependencies.
-//
-// Note: replace these inline IDs with the corresponding value from
-// `solana_sdk_ids` once the version is updated to 2.2.0.
-
-#[cfg(feature = "bincode")]
-const RECENT_BLOCKHASHES_ID: Pubkey =
-    Pubkey::from_str_const("SysvarRecentB1ockHashes11111111111111111111");
-
-#[cfg(feature = "bincode")]
-const RENT_ID: Pubkey = Pubkey::from_str_const("SysvarRent111111111111111111111111111111111");
-
-#[cfg(feature = "bincode")]
-#[cfg(test)]
-static_assertions::const_assert_eq!(solana_nonce::state::State::size(), NONCE_STATE_SIZE);
-/// The serialized size of the nonce state.
-#[cfg(feature = "bincode")]
-const NONCE_STATE_SIZE: usize = 80;
-
 #[cfg(test)]
 static_assertions::const_assert!(MAX_PERMITTED_DATA_LENGTH <= u32::MAX as u64);
 /// Maximum permitted size of account data (10 MiB).
@@ -49,27 +27,5 @@ pub const MAX_PERMITTED_ACCOUNTS_DATA_ALLOCATIONS_PER_TRANSACTION: i64 =
     MAX_PERMITTED_DATA_LENGTH as i64 * 2;
 
 pub mod program {
-    use solana_pubkey::declare_id;
-
-    declare_id!("11111111111111111111111111111111");
-}
-
-#[cfg(feature = "bincode")]
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use solana_program::sysvar::SysvarId;
-
-    #[allow(deprecated)]
-    #[test]
-    fn test_constants() {
-        // Ensure that the constants are in sync with the solana program.
-        assert_eq!(
-            RECENT_BLOCKHASHES_ID,
-            solana_program::sysvar::recent_blockhashes::RecentBlockhashes::id(),
-        );
-
-        // Ensure that the constants are in sync with the solana rent.
-        assert_eq!(RENT_ID, solana_program::sysvar::rent::Rent::id());
-    }
+    solana_pubkey::declare_id!("11111111111111111111111111111111");
 }
